chore: remove ARUseImprovedArtworkFilters feature flag
