+++ CHANGELOG.yml
@@ -9,6 +9,7 @@ upcoming:
     - Remove feature flag check for custom size filter - devon
     - Fix iOS app blank screen - ole, mounir
     - Replace "color" filter with "colors" filter - dzmitry tratsiak
+    - Remove ARUseImprovedArtworkFilters flag - iskounen
   user_facing:
     - Fix wrong input font on android - dzmitry
     - Fix home screen load failure on logout - brian
@@ -36,6 +37,7 @@ upcoming:
     - Fix time mismatch between auction list and individual auction page - matt
     - Fix swiping on filter options selects them - dzmitry tratsiak
     - Decode html entities in markdown - ole
+    - Release new price range filter to artwork grids - iskounen
 
 releases:
   - version: 6.9.2
