--- src/lib/Components/ArtworkFilter/Filters/SizeOptionsOld.tsx
@@ -1,54 +1,0 @@
-import { StackScreenProps } from "@react-navigation/stack"
-import { ArtworkFilterNavigationStack } from "lib/Components/ArtworkFilter"
-import {
-  AggregateOption,
-  aggregationForFilter,
-  FilterData,
-  FilterDisplayName,
-  FilterParamName,
-  ParamDefaultValues,
-} from "lib/Components/ArtworkFilter/ArtworkFilterHelpers"
-import { ArtworksFiltersStore, useSelectedOptionsDisplay } from "lib/Components/ArtworkFilter/ArtworkFilterStore"
-import React from "react"
-import { SingleSelectOptionScreen } from "./SingleSelectOption"
-
-interface SizeOptionsScreenProps extends StackScreenProps<ArtworkFilterNavigationStack, "SizeOptionsScreen"> {}
-
-export const SizeOptionsScreen: React.FC<SizeOptionsScreenProps> = ({ navigation }) => {
-  const paramName = FilterParamName.size
-  const aggregations = ArtworksFiltersStore.useStoreState((state) => state.aggregations)
-  const aggregation = aggregationForFilter(paramName, aggregations)
-
-  const selectFiltersAction = ArtworksFiltersStore.useStoreActions((state) => state.selectFiltersAction)
-
-  const options = aggregation?.counts.map((aggCount) => {
-    return {
-      displayText: aggCount.name,
-      paramName,
-      paramValue: aggCount.value,
-    }
-  })
-
-  const allOption: FilterData = { displayText: "All", paramName, paramValue: ParamDefaultValues.dimensionRange }
-  const displayOptions = [allOption].concat(options ?? [])
-  const selectedOptions = useSelectedOptionsDisplay()
-  const selectedOption = selectedOptions.find((option) => option.paramName === paramName)!
-
-  const selectOption = (option: AggregateOption) => {
-    selectFiltersAction({
-      displayText: option.displayText,
-      paramValue: option.paramValue,
-      paramName,
-    })
-  }
-
-  return (
-    <SingleSelectOptionScreen
-      onSelect={selectOption}
-      filterHeaderText={FilterDisplayName.size}
-      filterOptions={displayOptions}
-      selectedOption={selectedOption}
-      navigation={navigation}
-    />
-  )
-}
