+++ src/lib/store/config/features.ts
@@ -89,11 +89,6 @@ export const features = defineFeatures({
     description: "Enable custom share sheet",
     showInAdminMenu: true,
   },
-  ARUseImprovedArtworkFilters: {
-    readyForRelease: false,
-    description: "Use improved artwork filters",
-    showInAdminMenu: true,
-  },
   AREnableOrderHistoryOption: {
     readyForRelease: false,
     description: "Enable Order History in settings",
